Fix docker runtime not including required libraries
Additionally:
- extracts hardcoded versions into variables
- bumps rust version
- adds cache mounts for faster local builds

The runtime image, originally based on Alpine, was incompatible
with the built binaries. Critically, it was missing glibc, although
openssl could also have been an issue. To fix that, we're now
using a compatible, glibc-based, base image. 

I've picked Google's Distroless image as a good, small base.
You could also use Debian Slim, but that results in a significantly
larger image. There are other options that result in smaller
images, but this is one of the simplest - and the priority is to
have something that works.